create VenueOffers
